+++ .env.example
@@ -1,1 +1,54 @@
-# Example environment file
+
+# Security
+SECRET_KEY=change_this_to_random_secret_key
+JWT_SECRET_KEY=change_this_to_different_random_secret
+SESSION_COOKIE_SECURE=True
+
+# Database
+DATABASE_URL=sqlite:///farm_data.db
+# For PostgreSQL:
+# DATABASE_URL=postgresql://user:password@localhost/farmdb
+
+# Redis (for caching and Celery)
+REDIS_URL=redis://localhost:6379/0
+
+# API Configuration
+API_HOST=0.0.0.0
+API_PORT=5000
+API_DEBUG=False
+
+# Rate Limiting
+RATE_LIMIT_DEFAULT=100 per hour
+RATE_LIMIT_LOGIN=5 per minute
+
+# Hardware Configuration
+GPIO_MODE=BCM
+TEMPERATURE_SENSOR_PIN=4
+HUMIDITY_SENSOR_PIN=17
+LIGHT_SENSOR_I2C_ADDRESS=0x23
+
+# Monitoring
+PROMETHEUS_PORT=9090
+LOG_LEVEL=INFO
+LOG_FILE=/var/log/farm/app.log
+
+# Backup Configuration
+BACKUP_DIR=/home/pi/backups
+REMOTE_BACKUP_HOST=
+REMOTE_BACKUP_USER=
+REMOTE_BACKUP_DIR=
+
+# External Services (optional)
+WEATHER_API_KEY=
+MQTT_BROKER=
+MQTT_PORT=1883
+MQTT_USERNAME=
+MQTT_PASSWORD=
+
+# Email Notifications (optional)
+SMTP_HOST=
+SMTP_PORT=587
+SMTP_USERNAME=
+SMTP_PASSWORD=
+NOTIFICATION_EMAIL=
+
