+++ scripts/install_dependencies.py
@@ -1,285 +1,3 @@
-# Requirements and setup files for production deployment
-
-# === requirements.in ===
-# Core dependencies with pinned versions for production stability
-# Use pip-compile to generate requirements.txt with hashes
-
-# Core Framework
-Flask==2.3.3
-flask-cors==4.0.0
-flask-limiter==3.5.0
-flask-bcrypt==1.0.1
-flask-restx==1.2.0
-
-# Authentication & Security
-PyJWT==2.8.0
-python-dotenv==1.0.0
-cryptography==41.0.4
-
-# Database
-SQLAlchemy==2.0.21
-alembic==1.12.0
-redis==5.0.0
-
-# Hardware Interface (Raspberry Pi)
-RPi.GPIO==0.7.1
-adafruit-circuitpython-dht==4.0.2
-w1thermsensor==2.0.0
-smbus2==0.4.3
-
-# Monitoring & Logging
-prometheus-client==0.17.1
-python-json-logger==2.0.7
-
-# Task Queue
-celery==5.3.4
-celery[redis]==5.3.4
-
-# Data Processing
-pandas==2.1.1
-numpy==1.25.2
-
-# Testing
-pytest==7.4.2
-pytest-cov==4.1.0
-pytest-mock==3.11.1
-factory-boy==3.3.0
-
-# Development
-black==23.9.1
-flake8==6.1.0
-mypy==1.5.1
-pre-commit==3.4.0
-
-# === setup.py ===
-"""
-Setup configuration for OpenSourceControlledEnvironments
-"""
-
-from setuptools import setup, find_packages
-
-with open("README.md", "r", encoding="utf-8") as fh:
-    long_description = fh.read()
-
-setup(
-    name="opensourcecontrolledenvironments",
-    version="1.0.0",
-    author="HydroFarmerJason",
-    author_email="",
-    description="Open source controlled environment agriculture system",
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    url="https://github.com/HydroFarmerJason/OpenSourceControlledEnvironments",
-    packages=find_packages(where="src"),
-    package_dir={"": "src"},
-    classifiers=[
-        "Development Status :: 4 - Beta",
-        "Intended Audience :: Developers",
-        "Intended Audience :: Education",
-        "Intended Audience :: Science/Research",
-        "Topic :: Scientific/Engineering :: Bio-Informatics",
-        "License :: OSI Approved :: MIT License",
-        "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
-        "Operating System :: POSIX :: Linux",
-    ],
-    python_requires=">=3.9",
-    install_requires=[
-        # List core requirements here
-        # Full list in requirements.txt
-    ],
-    extras_require={
-        "dev": [
-            "pytest>=7.4.2",
-            "black>=23.9.1",
-            "flake8>=6.1.0",
-            "mypy>=1.5.1",
-        ],
-        "docs": [
-            "sphinx>=7.2.6",
-            "sphinx-rtd-theme>=1.3.0",
-        ],
-    },
-    entry_points={
-        "console_scripts": [
-            "farm-controller=farm.cli:main",
-            "farm-api=farm.api:run_server",
-        ],
-    },
-)
-
-# === pyproject.toml ===
-"""
-Modern Python project configuration
-"""
-
-[build-system]
-requires = ["setuptools>=68.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "opensourcecontrolledenvironments"
-version = "1.0.0"
-description = "Open source controlled environment agriculture system"
-readme = "README.md"
-requires-python = ">=3.9"
-license = {text = "MIT"}
-authors = [
-    {name = "HydroFarmerJason"},
-]
-maintainers = [
-    {name = "HydroFarmerJason"},
-]
-classifiers = [
-    "Development Status :: 4 - Beta",
-    "Intended Audience :: Developers",
-    "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-]
-
-[project.urls]
-Homepage = "https://github.com/HydroFarmerJason/OpenSourceControlledEnvironments"
-Documentation = "https://hydrofarmerjason.github.io/OpenSourceControlledEnvironments"
-Repository = "https://github.com/HydroFarmerJason/OpenSourceControlledEnvironments"
-"Bug Tracker" = "https://github.com/HydroFarmerJason/OpenSourceControlledEnvironments/issues"
-
-[tool.black]
-line-length = 100
-target-version = ['py39', 'py310', 'py311']
-include = '\.pyi?$'
-extend-exclude = '''
-/(
-  # directories
-  \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | build
-  | dist
-)/
-'''
-
-[tool.isort]
-profile = "black"
-line_length = 100
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
-
-[tool.mypy]
-python_version = "3.9"
-warn_return_any = true
-warn_unused_configs = true
-disallow_untyped_defs = true
-disallow_incomplete_defs = true
-check_untyped_defs = true
-disallow_untyped_decorators = true
-no_implicit_optional = true
-warn_redundant_casts = true
-warn_unused_ignores = true
-warn_no_return = true
-warn_return_any = true
-strict_equality = true
-
-[tool.pytest.ini_options]
-minversion = "7.0"
-addopts = [
-    "-ra",
-    "--strict-markers",
-    "--cov=src",
-    "--cov-branch",
-    "--cov-report=term-missing:skip-covered",
-    "--cov-report=html",
-    "--cov-report=xml",
-]
-testpaths = ["tests"]
-pythonpath = ["src"]
-
-[tool.coverage.run]
-source = ["src"]
-branch = true
-
-[tool.coverage.report]
-exclude_lines = [
-    "pragma: no cover",
-    "def __repr__",
-    "if self.debug:",
-    "if __name__ == .__main__.:",
-    "raise AssertionError",
-    "raise NotImplementedError",
-    "if 0:",
-    "if False:",
-]
-
-# === .env.example ===
-"""
-Environment configuration template
-Copy to .env and update with your values
-"""
-
-# Security
-SECRET_KEY=change_this_to_random_secret_key
-JWT_SECRET_KEY=change_this_to_different_random_secret
-SESSION_COOKIE_SECURE=True
-
-# Database
-DATABASE_URL=sqlite:///farm_data.db
-# For PostgreSQL:
-# DATABASE_URL=postgresql://user:password@localhost/farmdb
-
-# Redis (for caching and Celery)
-REDIS_URL=redis://localhost:6379/0
-
-# API Configuration
-API_HOST=0.0.0.0
-API_PORT=5000
-API_DEBUG=False
-
-# Rate Limiting
-RATE_LIMIT_DEFAULT=100 per hour
-RATE_LIMIT_LOGIN=5 per minute
-
-# Hardware Configuration
-GPIO_MODE=BCM
-TEMPERATURE_SENSOR_PIN=4
-HUMIDITY_SENSOR_PIN=17
-LIGHT_SENSOR_I2C_ADDRESS=0x23
-
-# Monitoring
-PROMETHEUS_PORT=9090
-LOG_LEVEL=INFO
-LOG_FILE=/var/log/farm/app.log
-
-# Backup Configuration
-BACKUP_DIR=/home/pi/backups
-REMOTE_BACKUP_HOST=
-REMOTE_BACKUP_USER=
-REMOTE_BACKUP_DIR=
-
-# External Services (optional)
-WEATHER_API_KEY=
-MQTT_BROKER=
-MQTT_PORT=1883
-MQTT_USERNAME=
-MQTT_PASSWORD=
-
-# Email Notifications (optional)
-SMTP_HOST=
-SMTP_PORT=587
-SMTP_USERNAME=
-SMTP_PASSWORD=
-NOTIFICATION_EMAIL=
-
-# === install_dependencies.py ===
 """
 Automated dependency installation and verification
 """
