+++ src/farm/__init__.py
@@ -1,0 +1,9 @@
+# === src/farm/__init__.py ===
+"""
+OpenSource Controlled Environments
+Production-ready controlled environment agriculture system
+"""
+
+__version__ = "1.0.0"
+__author__ = "HydroFarmerJason"
+
