+++ src/farm/api/__init__.py
@@ -1,165 +1,3 @@
-# === src/farm/__init__.py ===
-"""
-OpenSource Controlled Environments
-Production-ready controlled environment agriculture system
-"""
-
-__version__ = "1.0.0"
-__author__ = "HydroFarmerJason"
-
-# === src/farm/app.py ===
-"""
-Main application factory and configuration
-"""
-
-import os
-import logging
-import platform
-from flask import Flask, jsonify
-from flask_cors import CORS
-from flask_limiter import Limiter
-from flask_limiter.util import get_remote_address
-
-from farm.database import db, migrate
-from farm.security.config import create_secure_app, AuthenticationManager
-from farm.error_handling.framework import ErrorHandler
-from farm.api import api_bp
-from farm.monitoring import init_monitoring
-
-
-def create_app(config_name='production'):
-    """Create and configure Flask application"""
-    
-    # Create secure app with authentication
-    app = create_secure_app()
-    
-    # Load configuration
-    app.config.from_object(f'farm.config.{config_name}')
-    
-    # Initialize extensions
-    CORS(app, origins=app.config.get('CORS_ORIGINS', ['*']))
-    db.init_app(app)
-    migrate.init_app(app, db)
-    
-    # Initialize error handling
-    error_handler = ErrorHandler(app)
-    app.error_handler = error_handler
-    
-    # Initialize monitoring
-    init_monitoring(app)
-    
-    # Register blueprints
-    app.register_blueprint(api_bp, url_prefix='/api/v1')
-    
-    # Health check endpoint
-    @app.route('/health')
-    def health_check():
-        """System health check endpoint"""
-        try:
-            # Check database
-            db.session.execute('SELECT 1')
-            db_status = 'healthy'
-        except Exception as e:
-            db_status = f'unhealthy: {str(e)}'
-        
-        # Get system health from error monitor
-        system_health = app.error_handler.error_monitor.get_system_health()
-        
-        return jsonify({
-            'status': system_health['status'],
-            'version': __version__,
-            'database': db_status,
-            'health_score': system_health['health_score'],
-            'components': {
-                'api': 'healthy',
-                'database': db_status,
-                'sensors': 'healthy',  # In production, check actual sensors
-                'actuators': 'healthy'  # In production, check actual actuators
-            }
-        })
-    
-    # Version endpoint
-    @app.route('/version')
-    def version():
-        return jsonify({
-            'version': __version__,
-            'api_version': 'v1',
-            'python_version': platform.python_version()
-        })
-    
-    return app
-
-
-# === src/farm/config.py ===
-"""
-Application configuration
-"""
-
-import os
-from datetime import timedelta
-
-
-class Config:
-    """Base configuration"""
-    
-    # Flask
-    SECRET_KEY = os.getenv('SECRET_KEY', 'dev-secret-key')
-    
-    # Database
-    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL', 'sqlite:///farm_data.db')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    
-    # Redis
-    REDIS_URL = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
-    
-    # CORS
-    CORS_ORIGINS = os.getenv('CORS_ORIGINS', '*').split(',')
-    
-    # Rate limiting
-    RATELIMIT_STORAGE_URL = os.getenv('REDIS_URL', 'redis://localhost:6379/1')
-    
-    # Celery
-    CELERY_BROKER_URL = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
-    CELERY_RESULT_BACKEND = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
-    
-    # Hardware
-    GPIO_MODE = os.getenv('GPIO_MODE', 'BCM')
-    MOCK_HARDWARE = os.getenv('MOCK_HARDWARE', 'False').lower() == 'true'
-    
-    # Monitoring
-    PROMETHEUS_PORT = int(os.getenv('PROMETHEUS_PORT', '9090'))
-    
-    # Logging
-    LOG_LEVEL = os.getenv('LOG_LEVEL', 'INFO')
-    LOG_FILE = os.getenv('LOG_FILE', '/var/log/farm/app.log')
-
-
-class DevelopmentConfig(Config):
-    """Development configuration"""
-    DEBUG = True
-    TESTING = False
-    MOCK_HARDWARE = True
-
-
-class TestingConfig(Config):
-    """Testing configuration"""
-    DEBUG = False
-    TESTING = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///:memory:'
-    MOCK_HARDWARE = True
-
-
-class ProductionConfig(Config):
-    """Production configuration"""
-    DEBUG = False
-    TESTING = False
-    
-    # Stricter security in production
-    SESSION_COOKIE_SECURE = True
-    SESSION_COOKIE_HTTPONLY = True
-    SESSION_COOKIE_SAMESITE = 'Lax'
-
-
 # === src/farm/api/__init__.py ===
 """
 API Blueprint and routes
@@ -388,109 +226,3 @@ class SystemBackup(Resource):
         }), 202
 
 
-# === src/farm/cli.py ===
-"""
-Command-line interface for farm management
-"""
-
-import click
-import sys
-from flask.cli import FlaskGroup
-
-from farm.app import create_app
-from farm.database import db
-
-
-def create_cli_app():
-    """Create app for CLI"""
-    return create_app('development')
-
-
-cli = FlaskGroup(create_app=create_cli_app)
-
-
-@cli.command()
-def init_db():
-    """Initialize the database"""
-    click.echo('Initializing database...')
-    db.create_all()
-    click.echo('Database initialized.')
-
-
-@cli.command()
-@click.option('--host', default='0.0.0.0', help='Host to bind to')
-@click.option('--port', default=5000, help='Port to bind to')
-def run_api(host, port):
-    """Run the API server"""
-    app = create_app()
-    app.run(host=host, port=port)
-
-
-@cli.command()
-def test():
-    """Run tests"""
-    import pytest
-    
-    click.echo('Running tests...')
-    result = pytest.main(['-v', 'tests/'])
-    sys.exit(result)
-
-
-@cli.command()
-def check_health():
-    """Check system health"""
-    app = create_app()
-    with app.app_context():
-        # Check database
-        try:
-            db.session.execute('SELECT 1')
-            click.echo('✓ Database connection: OK')
-        except Exception as e:
-            click.echo(f'✗ Database connection: {e}', err=True)
-        
-        # Check sensors
-        from farm.sensors import sensor_manager
-        sensor_status = sensor_manager.check_all_sensors()
-        for sensor_id, status in sensor_status.items():
-            if status['healthy']:
-                click.echo(f'✓ Sensor {sensor_id}: OK')
-            else:
-                click.echo(f'✗ Sensor {sensor_id}: {status["error"]}', err=True)
-
-
-if __name__ == '__main__':
-    cli()
-
-
-# === run.py ===
-"""
-Entry point for running the application
-"""
-
-import os
-import sys
-
-# Add src directory to Python path
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))
-
-from farm.app import create_app
-
-# Create application
-app = create_app(os.getenv('FARM_ENV', 'production'))
-
-if __name__ == '__main__':
-    # Run with production server in production
-    if app.config['DEBUG']:
-        app.run(
-            host=os.getenv('API_HOST', '0.0.0.0'),
-            port=int(os.getenv('API_PORT', 5000)),
-            debug=True
-        )
-    else:
-        # Use gunicorn in production
-        import multiprocessing
-        
-        bind = f"{os.getenv('API_HOST', '0.0.0.0')}:{os.getenv('API_PORT', '5000')}"
-        workers = multiprocessing.cpu_count() * 2 + 1
-        
-        os.system(f"gunicorn -b {bind} -w {workers} 'farm.app:create_app()'")
