+++ src/farm/config.py
@@ -1,0 +1,70 @@
+# === src/farm/config.py ===
+"""
+Application configuration
+"""
+
+import os
+from datetime import timedelta
+
+
+class Config:
+    """Base configuration"""
+    
+    # Flask
+    SECRET_KEY = os.getenv('SECRET_KEY', 'dev-secret-key')
+    
+    # Database
+    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL', 'sqlite:///farm_data.db')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    
+    # Redis
+    REDIS_URL = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
+    
+    # CORS
+    CORS_ORIGINS = os.getenv('CORS_ORIGINS', '*').split(',')
+    
+    # Rate limiting
+    RATELIMIT_STORAGE_URL = os.getenv('REDIS_URL', 'redis://localhost:6379/1')
+    
+    # Celery
+    CELERY_BROKER_URL = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
+    CELERY_RESULT_BACKEND = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
+    
+    # Hardware
+    GPIO_MODE = os.getenv('GPIO_MODE', 'BCM')
+    MOCK_HARDWARE = os.getenv('MOCK_HARDWARE', 'False').lower() == 'true'
+    
+    # Monitoring
+    PROMETHEUS_PORT = int(os.getenv('PROMETHEUS_PORT', '9090'))
+    
+    # Logging
+    LOG_LEVEL = os.getenv('LOG_LEVEL', 'INFO')
+    LOG_FILE = os.getenv('LOG_FILE', '/var/log/farm/app.log')
+
+
+class DevelopmentConfig(Config):
+    """Development configuration"""
+    DEBUG = True
+    TESTING = False
+    MOCK_HARDWARE = True
+
+
+class TestingConfig(Config):
+    """Testing configuration"""
+    DEBUG = False
+    TESTING = True
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///:memory:'
+    MOCK_HARDWARE = True
+
+
+class ProductionConfig(Config):
+    """Production configuration"""
+    DEBUG = False
+    TESTING = False
+    
+    # Stricter security in production
+    SESSION_COOKIE_SECURE = True
+    SESSION_COOKIE_HTTPONLY = True
+    SESSION_COOKIE_SAMESITE = 'Lax'
+
+
